Moved one testcase from ScrollTBTest to spreadsheet module
ScrollTBTest.testHiddenColumnHeaderScrollingFix() test metod
is not related to charts. Moved it to vaadin spreadsheet module
also moved screenshot references and test file
