+++ vaadin-spreadsheet-charts/src/test/java/com/vaadin/spreadsheet/charts/interactiontests/ScrollTBTest.java
@@ -17,20 +17,6 @@ public class ScrollTBTest extends AbstractSpreadsheetTestCase {
         spreadsheetElement.scroll(850);
         compareScreen("overlayScroll");
     }
-
-    @Test
-    public void testHiddenColumnHeaderScrollingFix() throws IOException {
-    	headerPage.loadFile("hidden.xlsx",this);
-    	SpreadsheetElement spreadsheetElement = $(SpreadsheetElement.class).first();
-        spreadsheetElement.scroll(128);
-        spreadsheetElement.scroll(128);
-        spreadsheetElement.scroll(512);
-        spreadsheetElement.scroll(-256);
-        spreadsheetElement.scroll(512);
-        spreadsheetElement.scroll(-1024);
-        compareScreen("hiddenColumnScroll");
-    }
-
     @Override
     public Class<?> getUIClass() {
         return SpreadsheetChartsDemoUI.class;
